utils.ts file removed
