+++ template/src/routes/index.ts
@@ -2,6 +2,5 @@ export { RouteComponent } from './route-component';
 export { AppLink } from './app-link';
 export { AppRedirect } from './app-redirect';
 export { goToPage } from './route-helpers';
-export { routes } from './routes';
-export { RouteName } from './utils';
+export { routes, RouteName } from './routes';
 export { Router } from './router';
