+++ template/src/routes/route-component.ts
@@ -1,7 +1,7 @@
 import { Home } from 'pages/home';
 import { About } from 'pages/about';
 import { NotFound } from 'pages/not-found';
-import { RouteName } from './utils';
+import { RouteName } from './routes';
 
 const RouteComponent = {
   [RouteName.Home]: Home,
